Add extension class for manifest
