+++ src/neoxp/Validators/TokenBase.cs
@@ -8,11 +8,9 @@
 // Redistribution and use in source and binary forms with or without
 // modifications are permitted.
 using Neo;
-using Neo.Network.P2P;
 using Neo.Network.P2P.Payloads;
 using Neo.Persistence;
 using Neo.VM;
-using Neo.Wallets;
 using System.Numerics;
 
 namespace NeoExpress.Validators;
